aml_pipeline/backend/requirements.txt
# Core ML and Data Processing
numpy>=1.23.5,<2.0.0
pandas>=1.5.3,<2.0.0
scikit-learn>=1.1.3,<2.0.0
xgboost>=1.6.2,<2.0.0
concrete-ml>=1.0.0,<2.0.0
snapml==1.15.0
datatable==1.1.0
scikit-optimize==0.9.0
joblib>=1.2.0,<2.0.0

# Web Framework and API
flask>=2.2.3,<3.0.0
werkzeug>=2.2.3,<3.0.0
flask-cors>=3.0.10,<4.0.0
gunicorn>=20.1.0,<21.0.0
streamlit==1.24.0
requests>=2.28.2,<3.0.0

# Logging and Monitoring
wandb==0.15.4
python-dotenv>=1.0.0,<2.0.0

# Testing and Development
pytest==7.3.1
ipython==8.12.0

# Utilities
python-dateutil>=2.8.2,<3.0.0
pytz>=2023.3,<2024.0
six>=1.16.0,<2.0.0
typing-extensions>=4.4.0,<5.0.0

# Build and Package Management
setuptools>=65.6.3,<66.0.0
wheel>=0.38.4,<0.39.0
pip>=23.0.1,<24.0.0

# Additional Dependencies
cryptography>=41.0.0,<42.0.0
numba>=0.56.4,<0.57.0
torch>=1.13.1,<1.14.0
torchvision>=0.14.1,<0.15.0
boto3>=1.23.5,<2.0.0